requirements.txt
    Flask==2.3.1
    git+https://github.com/Omegup/pyzk.git
    Flask-SQLAlchemy==3.1.1
    netifaces==0.11.0
    scapy==2.5.0
    python-nmap==0.7.1
    wmi==1.5.1; sys_platform == 'win32'
    marshmallow==3.20.1